refactored actions
